+++ R/wrappers.R
@@ -6,6 +6,7 @@
 #' @param verbose whether to print messages
 #' @param downsample downsample when too many datapoints are present
 #' @param nsamp how many random sampling color combinations to test, default 50000
+#' @param polychrome_recolor whether to replace the original colors with polychrome creation
 #' @param seed sampling randomization seed
 #' @param autoswitch try to switch between colour and fill automatically
 #' @param layer layer to detect color, defaults to first
@@ -34,6 +35,7 @@ gg_color_repel <- function(g = ggplot2::last_plot(),
                            verbose = FALSE,
                            downsample = 5000,
                            nsamp = 50000,
+                           polychrome_recolor = FALSE,
                            seed = 34,
                            autoswitch = TRUE,
                            layer = 1,
@@ -53,6 +55,7 @@ gg_color_repel <- function(g = ggplot2::last_plot(),
     col = col, verbose = verbose,
     downsample = downsample,
     nsamp = nsamp, seed = seed,
+    polychrome_recolor = polychrome_recolor,
     sim = sim, severity = severity,
     autoswitch = autoswitch, layer = layer,
     out_orig = out_orig,